+++ wl_agent.cpp
@@ -32,7 +32,7 @@
 // player state info
 //
 boolean         running;
-long            thrustspeed;
+int32_t            thrustspeed;
 
 word plux,pluy;                 // player coordinates scaled to unsigned
 
@@ -57,7 +57,7 @@ statetype s_player = {false,0,0,(statefunc) T_Player,NULL,NULL};
 statetype s_attack = {false,0,0,(statefunc) T_Attack,NULL,NULL};
 
 
-long    playerxmove,playerymove;
+int32_t    playerxmove,playerymove;
 
 struct atkinf
 {
@@ -89,7 +89,7 @@ void SelectItem (void);
 boolean TryMove (objtype *ob);
 void T_Player (objtype *ob);
 
-void ClipMove (objtype *ob, long xmove, long ymove);
+void ClipMove (objtype *ob, int32_t xmove, int32_t ymove);
 
 /*
 =============================================================================
@@ -143,7 +143,7 @@ void CheckWeaponChange (void)
 
 void ControlMovement (objtype *ob)
 {
-        long    oldx,oldy;
+        int32_t    oldx,oldy;
         int             angle;//,maxxmove;
         int             angleunits;
 
@@ -327,7 +327,7 @@ void    UpdateFace (void)
 ===============
 */
 
-void    LatchNumber (int x, int y, int width, long number)
+void    LatchNumber (int x, int y, int width, int32_t number)
 {
         unsigned length,c;
         char    str[20];
@@ -513,7 +513,7 @@ void    DrawScore (void)
 ===============
 */
 
-void    GivePoints (long points)
+void    GivePoints (int32_t points)
 {
         gamestate.score += points;
         while (gamestate.score >= gamestate.nextextra)
@@ -792,7 +792,7 @@ boolean TryMove (objtype *ob)
 {
         int                     xl,yl,xh,yh,x,y;
         objtype         *check;
-        long            deltax,deltay;
+        int32_t            deltax,deltay;
 
         xl = (ob->x-PLAYERSIZE) >>TILESHIFT;
         yl = (ob->y-PLAYERSIZE) >>TILESHIFT;
@@ -878,9 +878,9 @@ boolean TryMove (objtype *ob)
 ===================
 */
 
-void ClipMove (objtype *ob, long xmove, long ymove)
+void ClipMove (objtype *ob, int32_t xmove, int32_t ymove)
 {
-        long    basex,basey;
+        int32_t    basex,basey;
 
         basex = ob->x;
         basey = ob->y;
@@ -892,8 +892,8 @@ void ClipMove (objtype *ob, long xmove, long ymove)
 
 #ifndef REMDEBUG
         if (noclip && ob->x > 2*TILEGLOBAL && ob->y > 2*TILEGLOBAL
-                        && ob->x < (((long)(mapwidth-1))<<TILESHIFT)
-                        && ob->y < (((long)(mapheight-1))<<TILESHIFT) )
+                        && ob->x < (((int32_t)(mapwidth-1))<<TILESHIFT)
+                        && ob->y < (((int32_t)(mapheight-1))<<TILESHIFT) )
                 return;         // walk through walls
 #endif
 
@@ -941,9 +941,9 @@ void VictoryTile (void)
 ===================
 */
 
-void Thrust (int angle, long speed)
+void Thrust (int angle, int32_t speed)
 {
-        long xmove,ymove;
+        int32_t xmove,ymove;
         unsigned offset;
 
 
@@ -1116,8 +1116,8 @@ void SpawnPlayer (int tilex, int tiley, int dir)
         player->tilex = tilex;
         player->tiley = tiley;
         player->areanumber = (byte) *(mapsegs[0]+(player->tiley<<mapshift)+player->tilex);
-        player->x = ((long)tilex<<TILESHIFT)+TILEGLOBAL/2;
-        player->y = ((long)tiley<<TILESHIFT)+TILEGLOBAL/2;
+        player->x = ((int32_t)tilex<<TILESHIFT)+TILEGLOBAL/2;
+        player->y = ((int32_t)tiley<<TILESHIFT)+TILEGLOBAL/2;
         player->state = &s_player;
         player->angle = (1-dir)*90;
         if (player->angle<0)
@@ -1144,7 +1144,7 @@ void SpawnPlayer (int tilex, int tiley, int dir)
 void    KnifeAttack (objtype *ob)
 {
         objtype *check,*closest;
-        long    dist;
+        int32_t    dist;
 
         SD_PlaySound (ATKKNIFESND);
 // actually fire
@@ -1181,7 +1181,7 @@ void    GunAttack (objtype *ob)
         objtype *check,*closest,*oldclosest;
         int             damage;
         int             dx,dy,dist;
-        long    viewdist;
+        int32_t    viewdist;
 
         switch (gamestate.weapon)
         {
@@ -1260,7 +1260,7 @@ void    GunAttack (objtype *ob)
 
 void VictorySpin (void)
 {
-        long    desty;
+        int32_t    desty;
 
         if (player->angle > 270)
         {
@@ -1275,7 +1275,7 @@ void VictorySpin (void)
                         player->angle = 270;
         }
 
-        desty = (((long)player->tiley-5)<<TILESHIFT)-0x3000;
+        desty = (((int32_t)player->tiley-5)<<TILESHIFT)-0x3000;
 
         if (player->y > desty)
         {
